BC-34 # refactor to prepare for more refs stuff
