+++ index.js
@@ -2,14 +2,19 @@
 
 // local modules
 
-const read = require('./lib/read');
-const write = require('./lib/write');
+const read = require('./lib/read.js');
+const refs = require('./lib/refs.js');
+const write = require('./lib/write.js');
 
 // this module
 
+/* ::
+export type JSONObject = { [id:string]: mixed }
+*/
+
 module.exports = {
-  findReferences: read.findReferences,
-  isFileReference: read.isFileReference,
+  findReferences: refs.findReferences,
+  isFileReference: refs.isFileReference,
   readData: read.readData,
   planWriteData: write.planWriteData,
   writeData: write.writeData,
