+++ lib/read.js
@@ -13,52 +13,13 @@ const traverse = require('traverse');
 
 // local modules
 
-const callbackify = require('./callbackify').callbackify;
+const callbackify = require('./callbackify.js').callbackify;
+const findReferences = require('./refs.js').findReferences;
+const isFileReference = require('./refs.js').isFileReference;
 
 // this module
 
-function isFileReference (
-  value /* : mixed */
-) /* : boolean */ {
-  return !!value && typeof value === 'object' && typeof value.$file === 'string';
-}
-
-/* ::
-type RefType = '$file'
-*/
-
-/* ::
-type Ref = {
-  path: Array<string>,
-  target: string,
-  type: RefType,
-  value?: string
-}
-*/
-
-/* ::
-type JSONObject = { [id:string]: mixed }
-*/
-
-function findReferences (
-  data /* : JSONObject */
-) /* : Promise<Array<Ref>> */ {
-  return new Promise((resolve, reject) => {
-    const refs = [];
-    /* eslint-disable no-invalid-this */ // traverse module's API
-    traverse(data).forEach(function () {
-      if (isFileReference(this.node)) {
-        refs.push({
-          path: this.path,
-          target: this.node.$file,
-          type: '$file'
-        });
-      }
-    });
-    /* eslint-enable no-invalid-this */
-    resolve(refs);
-  });
-}
+/* :: import type { Ref } from './refs.js' */
 
 function resolveReferences (
   refs /* : Array<Ref> */,
@@ -73,6 +34,8 @@ function resolveReferences (
     })
   );
 }
+
+/* :: import type { JSONObject } from '../index.js' */
 
 function replaceReferences (
   data /* : JSONObject */,
