+++ lib/write.js
@@ -16,8 +16,8 @@ const writeJson = require('write-json-file');
 
 // local modules
 
-const callbackify = require('./callbackify').callbackify;
-const findReferences = require('./read').findReferences;
+const callbackify = require('./callbackify.js').callbackify;
+const findReferences = require('./refs.js').findReferences;
 
 // this module
